+++ Nez.Portable/UI/Containers/Table.cs
@@ -164,7 +164,7 @@ namespace Nez.UI
 			if( _background == null )
 				return;
 			
-			_background.draw( graphics, x, y, width, height, new Color( color, color.A * parentAlpha ) );
+			_background.draw( graphics, x, y, width, height, new Color( color, (int)(color.A * parentAlpha) ) );
 		}
